don't autogenerate a debian_revision number
Idea is, packages we are building with an automatically generated
version number are all Debian natives packages (i.e. are using either
the "1.0" debian source format in "native mode" or using the "(3.0)
native" source format), and Debian natives packages must not (this is
now enforced by dpkg-source) have an debian revision. So remove it
from xivo-build-tools.